+++ lib/factories/abstract_validator.dart
@@ -5,8 +5,10 @@ abstract class AbstractValidator<E> {
   final List<AbstractRuleBuilder<E>> ruleBuilders = <AbstractRuleBuilder<E>>[];
 
   /// Creates a new builder to represent a list of rules.
-  AbstractRuleBuilder<E> ruleFor(dynamic Function(E e) func, {required String key}) {
-    final AbstractRuleBuilder<E> builder = AbstractRuleBuilder<E>(expression: func, key: key);
+  AbstractRuleBuilder<E> ruleFor(dynamic Function(E e) func,
+      {required String key}) {
+    final AbstractRuleBuilder<E> builder =
+        AbstractRuleBuilder<E>(expression: func, key: key);
     ruleBuilders.add(builder);
     return builder;
   }