+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [2.1.0] - [Breaking changes] - Nullability in validation error
+    * Make key, message, and code in validation errors required (default is empty string)
+
+## [2.0.1] - General maintanence
+    * Bump dependencies up to a more stable release
+    * Add a new regex expression matcher (isRegex)
+
 ## [2.0.0] - Null Safety
     * Update for null safety
     * [Breaking Change] - Add a key to rules to help identify the source of the error, this is useful for supplying the error message back to the correct form field. See example.